Add separate layer for message validation
