+++ src/main/java/com/rug/tno/App.java
@@ -23,6 +23,7 @@ public class App {
                         new FpPacketEncoder(),
                         new FpPayloadCodec(),
                         new DebugPrintLayer(),
+                        new FpValidationLayer(),
                         new FpSessionLayer(sessionManager),
                         new HlaForwardingLayer("localhost")
                 );
